Reorganized processor generators
